Use alternative GPIO pin in towel heater ESP8266
The ESP8266 was spontaneously turning on the towel heater upon boot /
wake up due to a short voltage spike in pin D8. Pin D6, unlike D8, is
not involved in the boot process, so it can be safely connected to a
pull-down resistor, which should keep the pin low even during boot.