Rework operating mode of towelheater device
Instead of running continuously during interesting time periods, wake the
device periodically just long enough to send sensor readings and act
upon commands.

For that to work, we use MQTT with retained messages and mark the switch
as optimistic. That way, command messages can be acted upon when the
device is active.

Alas, MQTT Discovery doesn't seem to support the 'retain' policy as part
of the configuration (or at least ESPHome isn't sending it, which
amounts to much of the same), so it required manual configuration on
Home Assistant as well.

A second less-than-great required change was explicitly unsetting the
command in the switch command topic once it's been acted upon.

Finally, the first value in the absolute humidity reading was returning
`nan` because it relied on unavailable-at-boot values, so a guard
against that was added.